+++ Source/GUI/WinCopies.GUI.Models.Shared/Models.cs
@@ -23,6 +23,7 @@ using System.Windows.Controls;
 using System.Windows.Controls.Primitives;
 using System.Windows.Input;
 using System.Windows.Media;
+using WinCopies.Commands;
 using WinCopies.Temp;
 
 #if WinCopies3