+++ Source/GUI/WinCopies.GUI.ViewModels.Shared/Controls/ViewModels.cs
@@ -19,7 +19,7 @@ using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Windows;
 using System.Windows.Input;
-
+using WinCopies.Commands;
 using WinCopies.GUI.Controls.Models;
 using WinCopies.Temp;
 using WinCopies.Util;