+++ Source/IO/WinCopies.IO.Shared/Process/ObjectModel/Copy.cs
@@ -34,14 +34,6 @@ using static WinCopies.IO.Resources.ExceptionMessages;
 
 namespace WinCopies.IO.Process.ObjectModel
 {
-    // TODO:
-    internal static class Win32
-    {
-        [DllImport(Microsoft.WindowsAPICodePack.NativeAPI.Consts.DllNames.Kernel32, SetLastError = true, CharSet = CharSet.Unicode)]
-        [return: MarshalAs(UnmanagedType.Bool)]
-        public static extern bool CreateDirectoryW([In, MarshalAs(UnmanagedType.LPWStr)] string lpPathName, [In] IntPtr lpSecurityAttributes);
-    }
-
     [ProcessGuid(ShellObjectInfo.Guids.ShellCopyProcessGuid)]
     public class CopyProcess<T> : ProcessObjectModelTypes<IPathInfo, T, ProcessError, ProcessDelegateTypes<IPathInfo, IProcessProgressDelegateParameter>.IProcessDelegates<ProcessDelegateTypes<IPathInfo, IProcessProgressDelegateParameter>.IProcessEventDelegates>, ProcessDelegateTypes<IPathInfo, IProcessProgressDelegateParameter>.IProcessEventDelegates, IProcessProgressDelegateParameter>.DefaultDestinationProcess where T : ProcessTypes<IPathInfo>.ProcessErrorTypes<ProcessError>.IProcessErrorFactories
     {
@@ -213,7 +205,7 @@ namespace WinCopies.IO.Process.ObjectModel
 
                 else
 
-                    result = Win32.CreateDirectoryW(destPath, IntPtr.Zero);
+                    result = Shell.CreateDirectoryW(destPath, IntPtr.Zero);
 
                 //void reportProgressCommon() => TryReportProgress((int)((_Paths.Count / InitialItemCount) * 100));